Handle BER/DER tags that are longer than one byte.
X.690 8.1.2 describes the encoding rules for tags including those that are
greater than (decimal) 30. The current code supports the format for values
up to 30 but mistakenly uses 31 (0x1f) as an "unknown" tag value in tests
while encoding it incorrectly and fails to support larger tags.

This change fixes both of these issues.

+++ src/ber/ber.rs
@@ -8,7 +8,7 @@ use oid::Oid;
 
 /// Defined in X.680 section 8.4
 #[derive(Clone, Copy, PartialEq, Eq)]
-pub struct BerTag(pub u8);
+pub struct BerTag(pub u32);
 
 newtype_enum!{
 impl debug BerTag {
